// src/store/category/ctaegory.enum.ts
enum TCategotyActionTypeEnum {
	GET_CATEGORY_SUCCESS = "GET_CATEGORY_SUCCESS",
	GET_CATEGORY_PENDING = "GET_CATEGORY_PENDING",
	GET_CATEGORY_FAILED = "GET_CATEGORY_FAILED",

	GET_CATEGORY_BY_ID_SUCCESS = "GET_CATEGORY_BY_ID_SUCCESS",
	GET_CATEGORY_BY_ID_PENDING = "GET_CATEGORY_BY_ID_PENDING",
	GET_CATEGORY_BY_ID_FAILED = "GET_CATEGORY_BY_ID_FAILED",
  
	ADD_CATEGORYS_PENDING = "ADD_CATEGORYS_PENDING",
	ADD_CATEGORYS_FAILED = "ADD_CATEGORYS_FAILED",
	ADD_CATEGORYS_SUCCESS = "ADD_CATEGORYS_SUCCESS",
  
	UPDATE_CATEGORYS_PENDING = "UPDATE_CATEGORYS_PENDING",
	UPDATE_CATEGORYS_FAILED = "UPDATE_CATEGORYS_FAILED",
	UPDATE_CATEGORYS_SUCCESS = "UPDATE_CATEGORYS_SUCCESS",
  
	DELETE_CATEGORYS_PENDING = "DELETE_CATEGORYS_PENDING",
	DELETE_CATEGORYS_FAILED = "DELETE_CATEGORYS_FAILED",
	DELETE_CATEGORYS_SUCCESS = "DELETE_CATEGORYS_SUCCESS",
}

export default TCategotyActionTypeEnum;